[#1529] Organize PrintBootReceipt support classes
- Clear naming: All classes now use BootReceipt* prefix for clarity
- Focused scope: Removed irrelevant retail receipt classes
- Better organization: Related functionality grouped logically
- Consistent placement: String extension moved to existing core_ext.rb
